semantic labeling stuff to its own file
