+++ libs/obs/src/CObservation3DRangeScan.cpp
@@ -21,7 +21,6 @@
 #include <mrpt/opengl/CPointCloud.h>
 #include <mrpt/poses/CPosePDF.h>
 #include <mrpt/serialization/CArchive.h>
-#include <mrpt/serialization/stl_serialization.h>
 #include <mrpt/system/CTimeLogger.h>
 #include <mrpt/system/filesystem.h>
 #include <mrpt/system/string_utils.h>
@@ -1199,99 +1198,6 @@ void CObservation3DRangeScan::getDescriptionAsText(std::ostream& o) const
 	  << relativePoseIntensityWRTDepth
 			 .getHomogeneousMatrixVal<CMatrixDouble44>()
 	  << endl;
-}
-
-using plib = CObservation3DRangeScan::TPixelLabelInfoBase;
-void plib::writeToStream(mrpt::serialization::CArchive& out) const
-{
-	const uint8_t version = 1;  // for possible future changes.
-	out << version;
-	// 1st: Save number MAX_NUM_DIFFERENT_LABELS so we can reconstruct the
-	// object in the class factory later on.
-	out << BITFIELD_BYTES;
-	// 2nd: data-specific serialization:
-	this->internal_writeToStream(out);
-}
-
-template <unsigned int BYTES_REQUIRED_>
-void CObservation3DRangeScan::TPixelLabelInfo<
-	BYTES_REQUIRED_>::internal_readFromStream(mrpt::serialization::CArchive& in)
-{
-	{
-		uint32_t nR, nC;
-		in >> nR >> nC;
-		pixelLabels.resize(nR, nC);
-		for (uint32_t c = 0; c < nC; c++)
-			for (uint32_t r = 0; r < nR; r++) in >> pixelLabels.coeffRef(r, c);
-	}
-	in >> pixelLabelNames;
-}
-template <unsigned int BYTES_REQUIRED_>
-void CObservation3DRangeScan::TPixelLabelInfo<BYTES_REQUIRED_>::
-	internal_writeToStream(mrpt::serialization::CArchive& out) const
-{
-	{
-		const auto nR = static_cast<uint32_t>(pixelLabels.rows());
-		const auto nC = static_cast<uint32_t>(pixelLabels.cols());
-		out << nR << nC;
-		for (uint32_t c = 0; c < nC; c++)
-			for (uint32_t r = 0; r < nR; r++) out << pixelLabels.coeff(r, c);
-	}
-	out << pixelLabelNames;
-}
-
-// Deserialization and class factory. All in one, ladies and gentlemen
-CObservation3DRangeScan::TPixelLabelInfoBase* plib::readAndBuildFromStream(
-	mrpt::serialization::CArchive& in)
-{
-	uint8_t version;
-	in >> version;
-
-	switch (version)
-	{
-		case 1:
-		{
-			// 1st: Read NUM BYTES
-			uint8_t bitfield_bytes;
-			in >> bitfield_bytes;
-
-			// Hand-made class factory. May be a good solution if there will be
-			// not too many different classes:
-			CObservation3DRangeScan::TPixelLabelInfoBase* new_obj = nullptr;
-			switch (bitfield_bytes)
-			{
-				case 1:
-					new_obj = new CObservation3DRangeScan::TPixelLabelInfo<1>();
-					break;
-				case 2:
-					new_obj = new CObservation3DRangeScan::TPixelLabelInfo<2>();
-					break;
-				case 3:
-				case 4:
-					new_obj = new CObservation3DRangeScan::TPixelLabelInfo<4>();
-					break;
-				case 5:
-				case 6:
-				case 7:
-				case 8:
-					new_obj = new CObservation3DRangeScan::TPixelLabelInfo<8>();
-					break;
-				default:
-					throw std::runtime_error(
-						"Unknown type of pixelLabel inner class while "
-						"deserializing!");
-			};
-			// 2nd: data-specific serialization:
-			new_obj->internal_readFromStream(in);
-
-			return new_obj;
-		}
-		break;
-
-		default:
-			MRPT_THROW_UNKNOWN_SERIALIZATION_VERSION(version);
-			break;
-	};
 }
 
 T3DPointsTo2DScanParams::T3DPointsTo2DScanParams()
